Adding new fields (#298)
* Adding new fields
New fields - contentManagedIdentity and contentPath
Adding these fields as a replacement to the sas url.

* Addressing some PR comments.

* Test Fix

* Adding switch - ExcludeArcMachines

* Putting ARC section removal inside hierarchy

* Updated unit tests with PR comments

* PR comments.

* Improved the output
Current code was leaving - {} in the array as the second element.

+++ source/Public/New-GuestConfigurationPolicy.ps1
@@ -31,6 +31,25 @@
         The public HTTP or HTTPS URI of the Guest Configuration package (.zip) to run via the created policy.
         Example: https://github.com/azure/auditservice/release/AuditService.zip
 
+        Note: If you are using an Azure storage account to store the custom machine configuration package artifact, you have two options for access:
+        1. Generate a blob shared access signature (SAS) token with read access and provide the full blob URI with the SAS token for the ContentUri parameter.
+        2. Create a user-assigned managed identity with read access to the storage account blob containing the package.
+            Provide the resource ID of the managed identity, a local path to the zipped package, a URI to the package without a SAS token and the ExcludeArcMachines parameter.
+            With this option, once the generated policy is applied, the managed identity will be used to download the package onto the target machine.
+
+    .PARAMETER ManagedIdentityResourceId
+        This is the identity that is used to download the package from storage account container instead of using SaS url.
+        The value for this parameter needs to be the resource id of the managed identity.
+        This is an option to use when the package is stored in a storage account and the storage account is protected by a managed identity.
+
+        Note: optional parameter. If this is specified, LocalContentPath and ExcludeArcMachines must also be specified.
+
+    .PARAMETER LocalContentPath
+        This is the path to the local package zip file. This is used to calculate the hash of the package.
+        The value of this parameter is not used in the policy definition.
+
+        Note: optional parameter. If this is specified, ManagedIdentityResourceId must also be specified.
+
     .PARAMETER ContentVersion
         If specified, the version of the Guest Configuration package (.zip) downloaded via the
         content URI must match this value.
@@ -95,6 +114,11 @@
     .PARAMETER Tag
         A hashtable of the tags that should be on machines to apply this policy on.
         If this is specified, the created policy will only be applied to machines with all the specified tags.
+
+
+    .PARAMETER ExcludeArcMachines
+        This parameter needs to be specified if the New-GuestConfigurationPolicy is using a User Assigned Identity.
+        Enabling this parameter will signal that users are aware of exclusion of Arc enabled servers in the definition.
 
     .EXAMPLE
         New-GuestConfigurationPolicy `
@@ -168,6 +192,14 @@ function New-GuestConfigurationPolicy
         [System.Uri]
         $ContentUri,
 
+        [Parameter(ParameterSetName='ManagedIdentity')]
+        [System.String]
+        $ManagedIdentityResourceId,
+
+        [Parameter(ParameterSetName='ManagedIdentity')]
+        [System.String]
+        $LocalContentPath,
+
         [Parameter()]
         [System.Version]
         $ContentVersion,
@@ -197,7 +229,11 @@ function New-GuestConfigurationPolicy
 
         [Parameter()]
         [System.Boolean]
-        $IncludeVMSS = $true
+        $IncludeVMSS = $true,
+
+        [Parameter()]
+        [Switch]
+        $ExcludeArcMachines
     )
 
     # Validate parameters
@@ -219,6 +255,11 @@ function New-GuestConfigurationPolicy
     if ($ContentUri.Scheme -notmatch '[http|https]')
     {
         throw "The specified package URI does not follow the HTTP or HTTPS scheme. Please specify a valid HTTP or HTTPS URI with the ContentUri parameter."
+    }
+
+    if ($PSCmdlet.ParameterSetName -eq 'ManagedIdentity' -and ([string]::IsNullOrWhiteSpace($ManagedIdentityResourceId) -or [string]::IsNullOrWhiteSpace($LocalContentPath)))
+    {
+        throw "Both ManagedIdentityResourceId and LocalContentPath must be provided together. Please include ManagedIdentityResourceId, LocalContentPath, and ExcludeArcMachines parameters."
     }
 
     $requiredParameterProperties = @('Name', 'DisplayName', 'Description', 'ResourceType', 'ResourceId', 'ResourcePropertyName')
@@ -275,16 +316,29 @@ function New-GuestConfigurationPolicy
 
     # Download package
     $tempPath = Reset-GCWorkerTempDirectory
+    $packagePath = Join-Path -Path $tempPath -ChildPath 'extracted'
 
-    $packageFileDownloadName = 'temp.zip'
-    $packageFileDownloadPath = Join-Path -Path $tempPath -ChildPath $packageFileDownloadName
-
-    if (Test-Path -Path $packageFileDownloadPath)
+    if (-not ([string]::IsNullOrWhiteSpace($ManagedIdentityResourceId) -or [string]::IsNullOrWhiteSpace($LocalContentPath)))
     {
-        $null = Remove-Item -Path $packageFileDownloadPath -Force
-    }
+        if (-not $ExcludeArcMachines)
+        {
+            throw "The ManagedIdentityResourceId and LocalContentPath parameters are defined but the -ExcludeArcMachines parameter is not. Managed identities cannot be used with Azure Arc machines. Please provide the -ExcludeArcMachines parameter to exclude Azure Arc machines and use a managed identity with this policy."
+        }
 
-    $null = Invoke-WebRequest -Uri $ContentUri -OutFile $packageFileDownloadPath
+        $packageFileDownloadPath = $LocalContentPath
+    }
+    else
+    {
+        $packageFileDownloadName = 'temp.zip'
+        $packageFileDownloadPath = Join-Path -Path $tempPath -ChildPath $packageFileDownloadName
+
+        if (Test-Path -Path $packageFileDownloadPath)
+        {
+            $null = Remove-Item -Path $packageFileDownloadPath -Force
+        }
+
+        $null = Invoke-WebRequest -Uri $ContentUri -OutFile $packageFileDownloadPath
+    }
 
     if ($null -eq (Get-Command -Name 'Get-FileHash' -ErrorAction 'SilentlyContinue'))
     {
@@ -293,7 +347,6 @@ function New-GuestConfigurationPolicy
     $contentHash = (Get-FileHash -Path $packageFileDownloadPath -Algorithm 'SHA256').Hash
 
     # Extract package
-    $packagePath = Join-Path -Path $tempPath -ChildPath 'extracted'
     $null = Expand-Archive -Path $packageFileDownloadPath -DestinationPath $packagePath -Force
 
     # Get configuration name
@@ -409,7 +462,13 @@ function New-GuestConfigurationPolicy
         Tag = $Tag
         IncludeVMSS = $IncludeVMSS
     }
-    $policyDefinitionContent = New-GuestConfigurationPolicyContent @policyDefinitionContentParameters
+
+    if (-not [string]::IsNullOrWhiteSpace($ManagedIdentityResourceId))
+    {
+        $policyDefinitionContentParameters.ManagedIdentityResourceId = $ManagedIdentityResourceId
+    }
+
+    $policyDefinitionContent = New-GuestConfigurationPolicyContent @policyDefinitionContentParameters -ExcludeArcMachines:$ExcludeArcMachines
 
     # Convert definition hashtable to JSON
     $policyDefinitionContentJson = (ConvertTo-Json -InputObject $policyDefinitionContent -Depth 100).Replace('\u0027', "'")
